# apps/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RESIDA API",
        "description": "REST API untuk sistem informasi manajemen warga perumahan. Mencakup manajemen penghuni, rumah, iuran, pembayaran, pengeluaran, dan laporan keuangan.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/residents": {
            "get": {
                "tags": [
                    "Residents"
                ],
                "summary": "Daftar penghuni (paginasi)",
                "description": "Mengembalikan daftar penghuni dengan paginasi. Mendukung pencarian (search), filter status hunian, dan sorting.",
                "operationId": "3a32e8e562ae44a95aa7f50edf35e12e",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari berdasarkan nama penghuni",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "resident_status",
                        "in": "query",
                        "description": "Filter status hunian (kontrak/tetap)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "kontrak",
                                "tetap"
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting. Prefix - untuk desc. Kolom: full_name, created_at, resident_status. Contoh: full_name, -created_at",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Data per halaman (max: 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data penghuni."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Residents"
                ],
                "summary": "Tambah penghuni baru",
                "description": "Menambahkan data penghuni baru beserta upload foto KTP. Foto KTP wajib diisi.",
                "operationId": "65baf45a21a18ad3f5c9eac6c83d527a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "ktp_photo",
                                    "resident_status",
                                    "phone_number",
                                    "marital_status"
                                ],
                                "properties": {
                                    "full_name": {
                                        "description": "Nama lengkap",
                                        "type": "string",
                                        "example": "John Doe",
                                        "maxLength": 255
                                    },
                                    "ktp_photo": {
                                        "description": "Foto KTP (JPG/JPEG/PNG, max 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "resident_status": {
                                        "type": "string",
                                        "example": "tetap",
                                        "enum": [
                                            "kontrak",
                                            "tetap"
                                        ]
                                    },
                                    "phone_number": {
                                        "description": "10-15 digit angka",
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "marital_status": {
                                        "type": "string",
                                        "example": "belum_menikah",
                                        "enum": [
                                            "sudah_menikah",
                                            "belum_menikah"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Penghuni berhasil ditambahkan."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "422": {
                        "description": "Validasi gagal."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/residents/{id}": {
            "get": {
                "tags": [
                    "Residents"
                ],
                "summary": "Detail penghuni",
                "description": "Detail penghuni berdasarkan ID, termasuk info rumah yang ditempati (jika ada).",
                "operationId": "9d3f9203667a1b59ebd9e5590e1007a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID penghuni",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data penghuni berhasil diambil."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Penghuni tidak ditemukan."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Residents"
                ],
                "summary": "Update data penghuni",
                "description": "Memperbarui data penghuni. Foto KTP opsional.",
                "operationId": "fe1e3fe63c48255ef32509ffcc6bbac6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID penghuni",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "resident_status",
                                    "phone_number",
                                    "marital_status"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "John Doe",
                                        "maxLength": 255
                                    },
                                    "ktp_photo": {
                                        "description": "Foto KTP baru (opsional)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "resident_status": {
                                        "type": "string",
                                        "example": "tetap",
                                        "enum": [
                                            "kontrak",
                                            "tetap"
                                        ]
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "marital_status": {
                                        "type": "string",
                                        "example": "belum_menikah",
                                        "enum": [
                                            "sudah_menikah",
                                            "belum_menikah"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data penghuni berhasil diperbarui."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Penghuni tidak ditemukan."
                    },
                    "422": {
                        "description": "Validasi gagal."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/residents/{id}/deactivate": {
            "patch": {
                "tags": [
                    "Residents"
                ],
                "summary": "Nonaktifkan penghuni (soft delete)",
                "description": "Nonaktifkan penghuni. Hanya bisa jika tidak memiliki hunian aktif.",
                "operationId": "b013b8dfe3b99d3540a4aa4458b97fea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID penghuni",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Penghuni berhasil dinonaktifkan."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Penghuni tidak ditemukan."
                    },
                    "409": {
                        "description": "Konflik - penghuni masih memiliki hunian aktif."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Sanctum. Format: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}